feat: json schema execute
