+++ src/lib/pages/execute/index.tsx
@@ -7,10 +7,10 @@ import { useInternalNavigate, useWasmConfig } from "lib/app-provider";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
 import { ContractSelectSection } from "lib/components/ContractSelectSection";
 import { CustomIcon } from "lib/components/icon";
-import { LoadingOverlay } from "lib/components/LoadingOverlay";
 import PageContainer from "lib/components/PageContainer";
-import { useExecuteCmds } from "lib/hooks";
+import { useSchemaStore } from "lib/providers/store";
 import { AmpTrackToExecute } from "lib/services/amplitude";
+import type { ContractDetail } from "lib/services/contractService";
 import type { ContractAddr } from "lib/types";
 import {
   getFirstQueryParam,
@@ -22,17 +22,32 @@ import {
 import { ExecuteArea } from "./components/ExecuteArea";
 
 const Execute = () => {
+  // ------------------------------------------//
+  // --------------DEPENDENCIES----------------//
+  // ------------------------------------------//
   useWasmConfig({ shouldRedirect: true });
+
   const router = useRouter();
   const navigate = useInternalNavigate();
+  const { getExecuteSchema } = useSchemaStore();
+
+  // ------------------------------------------//
+  // ------------------STATES------------------//
+  // ------------------------------------------//
   const [initialMsg, setInitialMsg] = useState("");
-  const [contractAddress, setContractAddress] = useState("");
+  const [contractAddress, setContractAddress] = useState("" as ContractAddr);
   const [initialFunds, setInitialFunds] = useState<Coin[]>([]);
+  const [codeHash, setCodeHash] = useState("");
+  const [codeId, setCodeId] = useState("");
 
-  const { isFetching, execCmds } = useExecuteCmds(
-    contractAddress as ContractAddr
-  );
+  // ------------------------------------------//
+  // ------------------LOGICS------------------//
+  // ------------------------------------------//
+  const schema = getExecuteSchema(codeHash);
 
+  // ------------------------------------------//
+  // ----------------CALLBACKS-----------------//
+  // ------------------------------------------//
   const goToQuery = () => {
     navigate({
       pathname: "/query",
@@ -53,6 +68,9 @@ const Execute = () => {
     [navigate]
   );
 
+  // ------------------------------------------//
+  // ---------------SIDE EFFECTS---------------//
+  // ------------------------------------------//
   useEffect(() => {
     const msgParam = getFirstQueryParam(router.query.msg);
     if (router.isReady) {
@@ -84,7 +102,6 @@ const Execute = () => {
 
   return (
     <PageContainer>
-      {isFetching && <LoadingOverlay />}
       <Flex mt={1} mb={8} justify="space-between">
         <Heading as="h5" variant="h5">
           Execute Contract
@@ -106,17 +123,24 @@ const Execute = () => {
         subtitle="You need to connect your wallet to perform this action"
         mb={8}
       />
+
       <ContractSelectSection
         mode="all-lists"
-        contractAddress={contractAddress as ContractAddr}
+        contractAddress={contractAddress}
         onContractSelect={onContractSelect}
+        successCallback={(data: ContractDetail) => {
+          setCodeHash(data.codeHash);
+          setCodeId(String(data.codeId));
+        }}
       />
 
       <ExecuteArea
-        contractAddress={contractAddress as ContractAddr}
+        contractAddress={contractAddress}
         initialMsg={initialMsg}
         initialFunds={initialFunds}
-        cmds={execCmds}
+        schema={schema}
+        codeId={codeId}
+        codeHash={codeHash}
       />
     </PageContainer>
   );
